+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 2.0.0
+
+### Changed
+
+- Replace shipping core package dependency.
+
+### Fixed
+
+- Include house number addition for Baden-Württemberg shipping addresses.
+
 ## 1.4.2
 
 ### Added